update motion rest and edit pong rules screen
A lot was done in this PR:

* MotionRest was updated

MotionRest::Controller was using Dispatch.once{} to initialize
singletons. The intention was to subclass MotionRest::Controller a bunch
of times (once for each controller/resource) and have a singleton made
for each of those subclasses. The problem is that Dispatch.once{}'s
entire purpose is to ensure the block is executed once per app run, at a
maximum. The Dispatch.once{} did not duplicate as you might think with
each subclass of MotionRest::Controller. Now we just have an ||=, which
is dangerous and should be done when NOTHING else asynchronous is going
on (e.g. early on in the app_delegate's on_load).

Additionally, we also have to keep a reference to all our singletons in
the app delegate so that they do not get deallocated. There may be a
smarter way to do this. For now, this is what I came up with. Look at
MotionRest::ApplicationControllers and the corresponding spec for
details.

* the edit pong rules screen was created

ProMotion-form was added to faciliate this. This is straight-forward.