Add @__action state to promotion screens (pollutes).
I’m okay with the pollution because this rails-like abstraction is already tightly coupled to ProMotion. Without promotion, this wouldn’t be possible (well, it would, but I’d have to write promotion!)

Typically in rails, when you render a form, you’re rendering the same template from the new action and the edit action. Rendering that form partial creates a form that “magically” knows its action (either create or update, respectively).

Adding an @__action state to the screen is the first step in allowing us to do similarly in our iOS counterpart. It creates a view variable that will know what action to return to and will be set on the screen by the controller when the screen is opened.

The two leading __underscores are not intended to hide the action state, but prevent naming collisions. If you want to have an @action variable in your view controller