Allow MotionRest::Controllers to call ProMotion’s open method.
ProMotion has two :open methods for opening screens (formerly termed iOS view controllers).

The first is in ProMotion::DelegateModule, which is a module that’s mixed into ProMotion::AppDelegate and opens the view first screen, including initializing the iOS window object on the application delegate.

The second is in ProMotion::ScreenNavigation and is mixed into all ProMotion::Screen instances.

I did not want to mix these modules into MotionRest::Controller and pollute it with a whole bunch of methods I don’t want in there.

Instead, a private :open method is defined on the MotionRest::Controller which determines who to forward the call to by checking to see if a window has been initialized. If a window has not been initialized, the :open call is forwarded to the application’s delegate. If a window has been initialized, the call is forwarded to the Screen object currently being displayed.